Get rid of acq_time variable showing up in level1c file
By:
Use pretty_formatting to avoid satazimuth_acq_time showing up.
Fixed bad use of scene.keys() so that acq_time is removed also form satzenith and sunzenith.
Also set include_lonlats=False to avoid that they are included again.
That did not help for current satpy master though.
A PR is on the way to satpy to fix it.